requirements.txt
streamlit>=1.27.0
openai>=1.0.0
firebase-admin>=6.2.0
pandas>=2.0.0
plotly>=5.17.0
python-docx>=0.8.11
pillow>=10.0.0
matplotlib>=3.7.0
pdfkit>=1.0.0
